refactor: reorganize and modularize components for better structure
