+++ src/app/fonts.ts
@@ -1,0 +1,14 @@
+import { DM_Sans, Playfair_Display } from "next/font/google";
+
+export const dmSans = DM_Sans({
+  subsets: ["latin"],
+  weight: ["300", "400", "500"],
+  variable: "--font-dm-sans",
+});
+
+export const playfair = Playfair_Display({
+  subsets: ["latin"],
+  weight: ["700", "900"],
+  style: ["normal", "italic"],
+  variable: "--font-playfair",
+});
